refactor: use mason-nvim-dap.nvim to manage dap servers.
