+++ lua/modules/plugins/tool.lua
@@ -111,6 +111,9 @@ tool["mfussenegger/nvim-dap"] = {
 			"rcarriga/nvim-dap-ui",
 			config = require("tool.dap.dapui"),
 		},
+		{
+			"jay-babu/mason-nvim-dap.nvim",
+		},
 	},
 }
